feat: preserve bundle for non-synthetic reloads
Feels a bit hacky but it works. When chokidar receives a file change event, set a cookie in the puppeteer instance before firing page.reload. Then in the framenavigation event, check for the cookie and if it exists, trigger a fresh rollup build and save the result to a cached build result var.